+++ src/pdb_file.h
@@ -34,7 +34,7 @@ class PDBFile
 
 		bool		load(const char* _filename);
 		void		getSymbolByAddress(uint64_t _address, rdebug::StackFrame& _frame);
-		uintptr_t	getSymbolID(uint64_t _address);
+		uint64_t	getSymbolID(uint64_t _address);
 		void		close();
 
 	private:
